Move console integration to a feature class
